+++ CHANGELOG.rst
@@ -6,6 +6,9 @@ Version 0.2.0
 
 *In development*
 
+- Values in :class:`~relief.Form` structures can now be validated by
+  subclassing :class:`~relief.Form` and adding `validate_{key}` methods.
+
 Version 0.1.0
 -------------